vexon-os/api/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
celery>=5.3.6
redis>=5.0.1
qdrant-client>=1.7.3
groq>=0.4.2
google-generativeai>=0.3.2
anthropic>=0.19.1
openai>=1.12.0
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.25
alembic>=1.13.1
firebase-admin>=6.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pydantic-settings>=2.1.0
requests>=2.31.0
websockets>=12.0
python-dotenv>=1.0.0
watchdog>=4.0.0
duckduckgo-search>=5.0.0
httpx>=0.27.0
